feat: add hook for fetching user sol data
